rename Commands directory to Console
